divided into components
