+++ src/pages/Home.jsx
@@ -1,30 +1,11 @@
-import React, { useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import axios from "axios";
-import {
-  Box,
-  TextField,
-  Grid,
-  Typography,
-  Container,
-  MenuItem,
-  FormControl,
-  InputLabel,
-  Select,
-  Slider,
-} from "@mui/material";
+import { Container } from "@mui/material";
 import TrendingMovies from "../components/features/home/TrendingMovies";
-import MovieCard from "../components/features/home/MovieCard";
+import MovieList from "../components/features/home/MovieList";
+import SearchFilters from "../components/features/home/SearchFilters";
 
 const API_KEY = import.meta.env.VITE_API_KEY;
-
-const genreOptions = [
-  { id: 28, name: "Action" },
-  { id: 35, name: "Comedy" },
-  { id: 18, name: "Drama" },
-  { id: 27, name: "Horror" },
-  { id: 10749, name: "Romance" },
-  // Add more as needed
-];
 
 function Home() {
   const [search, setSearch] = useState("");
@@ -67,71 +48,19 @@ function Home() {
 
   return (
     <Container maxWidth="lg">
-      {/* Filters */}
-      <Box sx={{ mt: 4, display: "flex", gap: 2, flexWrap: "wrap" }}>
-        <TextField
-          label="Search for movies..."
-          variant="outlined"
-          value={search}
-          onChange={(e) => {
-            const value = e.target.value;
-            setSearch(value);
-            localStorage.setItem("lastSearch", value);
-          }}
-        />
+      <SearchFilters
+        search={search}
+        setSearch={setSearch}
+        genre={genre}
+        setGenre={setGenre}
+        year={year}
+        setYear={setYear}
+        rating={rating}
+        setRating={setRating}
+      />
 
-        <FormControl sx={{ minWidth: 120 }}>
-          <InputLabel>Genre</InputLabel>
-          <Select
-            value={genre}
-            label="Genre"
-            onChange={(e) => setGenre(e.target.value)}
-          >
-            <MenuItem value="">All</MenuItem>
-            {genreOptions.map((g) => (
-              <MenuItem key={g.id} value={g.id}>
-                {g.name}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
+      <MovieList movies={movies} />
 
-        <TextField
-          label="Year"
-          type="number"
-          inputProps={{ min: 1900, max: new Date().getFullYear() }}
-          value={year}
-          onChange={(e) => setYear(e.target.value)}
-        />
-
-        <Box sx={{ width: 200 }}>
-          <Typography gutterBottom>Min Rating</Typography>
-          <Slider
-            value={rating}
-            onChange={(e, newVal) => setRating(newVal)}
-            min={0}
-            max={10}
-            step={0.5}
-            valueLabelDisplay="auto"
-          />
-        </Box>
-      </Box>
-
-      {/* Results */}
-      {(search || genre || year || rating > 0) && (
-        <Box sx={{ mt: 4 }}>
-          <Typography variant="h5" gutterBottom>
-            Search Results
-          </Typography>
-          <Grid container spacing={3}>
-            {movies.map((movie) => (
-              <MovieCard key={movie.id} movie={movie} />
-            ))}
-          </Grid>
-        </Box>
-      )}
-
-      {/* Trending */}
       <TrendingMovies />
     </Container>
   );
